// angular/orcamento/src/app/orcamento/planilha-vendas-hibrida/models/ficha-ceo.ts
export interface FichaCEO {
  idOrcamentoCenario: number;
  idOrcamento: number;
  nomeOrcamentoCenario: string;
  descricaoOrcamentoCenario: string;
  contatoCliente: string;
  concorrentes: string;
  pontosAtencao: string;
  prazoExecucaoProjeto: number;
  prazoExecucaoObra: number;
  validadeProposta: number;
}